[NativeSql][cpp] test roundrobin

+++ oap-native-sql/core/src/main/scala/org/apache/spark/shuffle/ColumnarShuffleWriter.scala
@@ -59,7 +59,7 @@ class ColumnarShuffleWriter[K, V](
   private val compressionEnabled = conf.getBoolean("spark.shuffle.compress", true)
   private val compressionCodec = conf.get("spark.io.compression.codec", "lz4")
   private val nativeBufferSize =
-    conf.getLong("spark.sql.execution.arrow.maxRecordsPerBatch", 4096)
+    conf.getInt("spark.sql.execution.arrow.maxRecordsPerBatch", 4096)
 
   private val jniWrapper = new ShuffleSplitterJniWrapper()
